Размер в SortableFSItem типа Long.

+++ app/src/main/java/com/github/aakumykov/extensible_sorting_comparator/fs_items/FileItem.kt
@@ -1,3 +1,3 @@
 package com.github.aakumykov.extensible_sorting_comparator.fs_items
 
-class FileItem(name: String, size: Int) : SimpleFSItem(isDir = false, name = name, size = size)
+class FileItem(name: String, size: Long) : SimpleFSItem(isDir = false, name = name, size = size)
