+++ app/src/main/java/com/github/aakumykov/extensible_sorting_comparator/fs_items_comparators/SortableFSItem.kt
@@ -3,6 +3,6 @@ package com.github.aakumykov.extensible_sorting_comparator.fs_items_comparators
 interface SortableFSItem {
     val isDir: Boolean
     val name: String
-    val size: Int
+    val size: Long
     val time: Long
 }
